a  fix

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,17 +6,17 @@
  */
 int main(void)
 {
-	unsigned int i, n, t1, t2, nextTerm;
+	unsigned long int i, n, t1, t2, nextTerm;
 
 	t1 = 1, t2 = 2, n = 98;
 
-	printf("%u", t1);
-	printf(", %u", t2);
+	printf("%lu", t1);
+	printf(", %lu", t2);
 
 	for (i = 3; i <= n; i++)
 	{
 		nextTerm = t1 + t2;
-		printf(", %u", nextTerm);
+		printf(", %lu", nextTerm);
 		t1 = t2;
 		t2 = nextTerm;
 	}
